+++ ClassicLootManager/Modules/Common/Changelog/Changelog.lua
@@ -7,7 +7,7 @@ CLM.ChangelogData = {
                 name = "Fixes",
                 data = {
                     { header = "Import", body = "Import should no longer sync improperly due to missing note." },
-                    { header = "Raid", body = "Raid names are now limited. |cfff48cbaZalán|r." },
+                    { header = "Raid", body = "Raid names are now limited to 128 characters. |cfff48cbaZalán|r." },
                 },
             },
         },